Impruve run time of unit tests
